+++ dotnetcore/WordAnalyser/Model/BookException.cs
@@ -6,7 +6,7 @@ namespace WordAnalyser.Model
 {
     public class BookException
     {
-        public int BookExceptionId { get; set; }
+        public Int64 BookExceptionId { get; set; }
         public string Top1020 { get; set; }
         public string ErrorMsg { get; set; }
         public string BookInfo { get; set; }
