+++ dotnetcore/WordAnalyser/Model/Word.cs
@@ -7,8 +7,8 @@ namespace WordAnalyser.Model
     public class Word
     {
         public Int64 WordId { get; set; }
-        public int LanguageType { get; set; }
+        public Int64 LanguageType { get; set; }
         public string WordBrief { get; set; }
-        public int WordUnicode { get; set; }
+        public Int64 WordUnicode { get; set; }
     }
 }
